[Update]: Implementing services. [ArtyomRusak]

+++ AR.EPAM/AR.EPAM.Core/Entities/Auction/Bid.cs
@@ -13,7 +13,7 @@ namespace AR.EPAM.Core.Entities.Auction
         public User User { get; set; }
         public int UserId { get; set; }
         public Lot Lot { get; set; }
-        public int LotId { get; set; }
+        public long LotId { get; set; }
         public Currency Currency { get; set; }
         public int CurrencyId { get; set; }
     }
